@@ -5,20 +5,19 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/http"
+	"net/url"
 	"os"
 	"regexp"
 	"strings"
 
 	"github.com/pkg/errors"
+	"github.com/projectdiscovery/nuclei/v2/pkg/catalog/disk"
 	"github.com/projectdiscovery/nvd"
 	"github.com/projectdiscovery/sliceutil"
 	"github.com/projectdiscovery/stringsutil"
 	"gopkg.in/yaml.v3"
-
-	"github.com/projectdiscovery/nuclei/v2/pkg/catalog"
 )
 
 const (
@@ -51,7 +50,7 @@ func main() {
 }
 
 func process() error {
-	tempDir, err := ioutil.TempDir("", "nuclei-nvd-%s")
+	tempDir, err := os.MkdirTemp("", "nuclei-nvd-%s")
 	if err != nil {
 		return err
 	}
@@ -61,7 +60,7 @@ func process() error {
 	if err != nil {
 		return err
 	}
-	catalog := catalog.New(*templateDir)
+	catalog := disk.NewCatalog(*templateDir)
 
 	paths, err := catalog.GetTemplatePath(*input)
 	if err != nil {
@@ -72,7 +71,16 @@ func process() error {
 		if err != nil {
 			return err
 		}
-		getCVEData(client, path, string(data))
+		dataString := string(data)
+
+		// First try to resolve references to tags
+		dataString, err = parseAndAddReferenceBasedTags(path, dataString)
+		if err != nil {
+			log.Printf("Could not parse reference tags %s: %s\n", path, err)
+			continue
+		}
+		// Next try and fill CVE data
+		getCVEData(client, path, dataString)
 	}
 	return nil
 }
@@ -98,6 +106,7 @@ var badRefs = []string{
 	"correkt.horse/",
 	"poc.wgpsec.org/",
 	"ctf-writeup.revers3c.com/",
+	"secunia.com/",
 }
 
 func getCVEData(client *nvd.Client, filePath, data string) {
@@ -229,7 +238,7 @@ func getCVEData(client *nvd.Client, filePath, data string) {
 
 	newTemplate := strings.ReplaceAll(data, infoBlockClean, newInfoBlockData)
 	if changed {
-		_ = ioutil.WriteFile(filePath, []byte(newTemplate), 0644)
+		_ = os.WriteFile(filePath, []byte(newTemplate), 0644)
 		fmt.Printf("Wrote updated template to %s\n", filePath)
 	}
 }
@@ -291,12 +300,15 @@ func parseAndAddCISAKevTagTemplate(path string, data string) (string, error) {
 		return "", errors.Wrap(err, "could not decode template yaml")
 	}
 	splitted := strings.Split(block.Info.Tags, ",")
+	if len(splitted) == 0 {
+		return data, nil
+	}
 
 	var cisaIndex = -1
 	for i, tag := range splitted {
 		// If we already have tag, return
 		if tag == "kev" {
-			return "", nil
+			return data, nil
 		}
 		if tag == "cisa" {
 			cisaIndex = i
@@ -307,9 +319,67 @@ func parseAndAddCISAKevTagTemplate(path string, data string) (string, error) {
 		splitted = append(splitted[:cisaIndex], splitted[cisaIndex+1:]...)
 	}
 	splitted = append(splitted, "kev")
-	final := strings.Join(splitted, ",")
-	replaced := strings.Replace(data, block.Info.Tags, final, -1)
-	return replaced, ioutil.WriteFile(path, []byte(replaced), os.ModePerm)
+	replaced := strings.ReplaceAll(data, block.Info.Tags, strings.Join(splitted, ","))
+	return replaced, os.WriteFile(path, []byte(replaced), os.ModePerm)
+}
+
+// parseAndAddReferenceBasedTags parses and adds reference based tags to templates
+func parseAndAddReferenceBasedTags(path string, data string) (string, error) {
+	block := &InfoBlock{}
+	if err := yaml.NewDecoder(strings.NewReader(data)).Decode(block); err != nil {
+		return "", errors.Wrap(err, "could not decode template yaml")
+	}
+	splitted := strings.Split(block.Info.Tags, ",")
+	if len(splitted) == 0 {
+		return data, nil
+	}
+	tagsCurrent := fmt.Sprintf("tags: %s", block.Info.Tags)
+	newTags := suggestTagsBasedOnReference(block.Info.Reference, splitted)
+
+	if len(newTags) == len(splitted) {
+		return data, nil
+	}
+	replaced := strings.ReplaceAll(data, tagsCurrent, fmt.Sprintf("tags: %s", strings.Join(newTags, ",")))
+	return replaced, os.WriteFile(path, []byte(replaced), os.ModePerm)
+}
+
+var referenceMapping = map[string]string{
+	"huntr.dev":               "huntr",
+	"hackerone.com":           "hackerone",
+	"tenable.com":             "tenable",
+	"packetstormsecurity.org": "packetstorm",
+	"seclists.org":            "seclists",
+	"wpscan.com":              "wpscan",
+	"packetstormsecurity.com": "packetstorm",
+	"exploit-db.com":          "edb",
+	"https://github.com/rapid7/metasploit-framework/": "msf",
+	"https://github.com/vulhub/vulhub/":               "vulhub",
+}
+
+func suggestTagsBasedOnReference(references, currentTags []string) []string {
+	uniqueTags := make(map[string]struct{})
+	for _, value := range currentTags {
+		uniqueTags[value] = struct{}{}
+	}
+
+	for _, reference := range references {
+		parsed, err := url.Parse(reference)
+		if err != nil {
+			continue
+		}
+		hostname := parsed.Hostname()
+
+		for value, tag := range referenceMapping {
+			if strings.HasSuffix(hostname, value) || strings.HasPrefix(reference, value) {
+				uniqueTags[tag] = struct{}{}
+			}
+		}
+	}
+	newTags := make([]string, 0, len(uniqueTags))
+	for tag := range uniqueTags {
+		newTags = append(newTags, tag)
+	}
+	return newTags
 }
 
 // Cloning struct from nuclei as we don't want any validation
