@@ -19,20 +19,21 @@ var (
 	failed  = aurora.Red("[✘]").String()
 
 	protocolTests = map[string]map[string]testutils.TestCase{
-		"http":          httpTestcases,
-		"network":       networkTestcases,
-		"dns":           dnsTestCases,
-		"workflow":      workflowTestcases,
-		"loader":        loaderTestcases,
-		"websocket":     websocketTestCases,
-		"headless":      headlessTestcases,
-		"whois":         whoisTestCases,
-		"ssl":           sslTestcases,
-		"code":          codeTestcases,
-		"templatesPath": templatesPathTestCases,
-		"templatesDir":  templatesDirTestCases,
-		"file":          fileTestcases,
-		"offlineHttp":   offlineHttpTestcases,
+		"http":            httpTestcases,
+		"network":         networkTestcases,
+		"dns":             dnsTestCases,
+		"workflow":        workflowTestcases,
+		"loader":          loaderTestcases,
+		"websocket":       websocketTestCases,
+		"headless":        headlessTestcases,
+		"whois":           whoisTestCases,
+		"ssl":             sslTestcases,
+		"code":            codeTestcases,
+		"templatesPath":   templatesPathTestCases,
+		"templatesDir":    templatesDirTestCases,
+		"file":            fileTestcases,
+		"offlineHttp":     offlineHttpTestcases,
+		"customConfigDir": customConfigDirTestCases,
 	}
 )
 
