@@ -1,6 +1,8 @@
 package filter
 
-import "github.com/projectdiscovery/nuclei/v2/pkg/catalog"
+import (
+	"github.com/projectdiscovery/nuclei/v2/pkg/catalog"
+)
 
 // PathFilter is a path based template filter
 type PathFilter struct {
@@ -15,7 +17,7 @@ type PathFilterConfig struct {
 }
 
 // NewPathFilter creates a new path filter from provided config
-func NewPathFilter(config *PathFilterConfig, catalogClient *catalog.Catalog) *PathFilter {
+func NewPathFilter(config *PathFilterConfig, catalogClient catalog.Catalog) *PathFilter {
 	filter := &PathFilter{
 		excludedTemplates:          catalogClient.GetTemplatesPath(config.ExcludedTemplates),
 		alwaysIncludedTemplatesMap: make(map[string]struct{}),
