@@ -43,9 +43,9 @@ func (e *Extractor) CompileExtractors() error {
 	}
 
 	for _, dslExp := range e.DSL {
-		compiled, err := govaluate.NewEvaluableExpressionWithFunctions(dslExp, dsl.HelperFunctions())
+		compiled, err := govaluate.NewEvaluableExpressionWithFunctions(dslExp, dsl.HelperFunctions)
 		if err != nil {
-			return fmt.Errorf("could not compile dsl: %s", dslExp)
+			return &dsl.CompilationError{DslSignature: dslExp, WrappedError: err}
 		}
 		e.dslCompiled = append(e.dslCompiled, compiled)
 	}
