@@ -11,11 +11,12 @@ import (
 // PayloadGenerator is the generator struct for generating payloads
 type PayloadGenerator struct {
 	Type     AttackType
+	catalog  catalog.Catalog
 	payloads map[string][]string
 }
 
 // New creates a new generator structure for payload generation
-func New(payloads map[string]interface{}, attackType AttackType, templatePath string, catalog *catalog.Catalog) (*PayloadGenerator, error) {
+func New(payloads map[string]interface{}, attackType AttackType, templatePath string, catalog catalog.Catalog) (*PayloadGenerator, error) {
 	if attackType.String() == "" {
 		attackType = BatteringRamAttack
 	}
@@ -36,12 +37,12 @@ func New(payloads map[string]interface{}, attackType AttackType, templatePath st
 		}
 	}
 
-	generator := &PayloadGenerator{}
+	generator := &PayloadGenerator{catalog: catalog}
 	if err := generator.validate(payloadsFinal, templatePath); err != nil {
 		return nil, err
 	}
 
-	compiled, err := loadPayloads(payloadsFinal)
+	compiled, err := generator.loadPayloads(payloadsFinal)
 	if err != nil {
 		return nil, err
 	}
